+++ common/src/main/java/net/ixdarklord/ultimine_addition/common/event/EventHandler.java
@@ -2,6 +2,7 @@ package net.ixdarklord.ultimine_addition.common.event;
 
 public class EventHandler {
     public static void register() {
+        DevEvents.init();
         MSCEvents.init();
         ChallengesEvents.init();
         IneligibleBlocksEvents.init();