+++ common/src/main/java/net/ixdarklord/ultimine_addition/core/CommonSetup.java
@@ -1,6 +1,6 @@
 package net.ixdarklord.ultimine_addition.core;
 
-import dev.ftb.mods.ftbultimine.integration.FTBUltiminePlugin;
+import dev.ftb.mods.ftbultimine.api.restriction.RegisterRestrictionHandlerEvent;
 import net.ixdarklord.ultimine_addition.api.CustomMSCApi;
 import net.ixdarklord.ultimine_addition.common.event.EventHandler;
 import net.ixdarklord.ultimine_addition.network.PayloadHandler;
@@ -8,7 +8,7 @@ import net.ixdarklord.ultimine_addition.config.ConfigHandler;
 
 public class CommonSetup {
     public static void init() {
-        FTBUltiminePlugin.register(new FTBUltimineIntegration());
+        RegisterRestrictionHandlerEvent.REGISTER.register(registry -> registry.register(FTBUltimineIntegration.INSTANCE));
         CustomMSCApi.init();
         ConfigHandler.register();
         Registration.register();