+++ common/src/main/java/net/ixdarklord/ultimine_addition/core/ShapeRegistryAccessor.java
@@ -1,0 +1,12 @@
+package net.ixdarklord.ultimine_addition.core;
+
+import dev.ftb.mods.ftbultimine.api.shape.Shape;
+import org.jetbrains.annotations.ApiStatus;
+
+import java.util.List;
+
+@ApiStatus.Internal
+public interface ShapeRegistryAccessor {
+    List<Shape> getShapesList();
+    Shape getDefaultShape();
+}