+++ common/src/main/java/net/ixdarklord/ultimine_addition/network/payloads/UpdateItemShapePayload.java
@@ -1,7 +1,7 @@
 package net.ixdarklord.ultimine_addition.network.payloads;
 
 import dev.architectury.networking.NetworkManager;
-import dev.ftb.mods.ftbultimine.shape.Shape;
+import dev.ftb.mods.ftbultimine.api.shape.Shape;
 import io.netty.buffer.ByteBuf;
 import net.ixdarklord.ultimine_addition.common.data.item.SelectedShapeData;
 import net.ixdarklord.ultimine_addition.common.menu.ShapeSelectorMenu;
@@ -11,6 +11,7 @@ import net.ixdarklord.ultimine_addition.core.Registration;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
@@ -33,7 +34,7 @@ public record UpdateItemShapePayload(String shapeId) implements CustomPacketPayl
                 return;
             }
 
-            Shape shape = FTBUltimineIntegration.getShape(message.shapeId);
+            Shape shape = FTBUltimineIntegration.getShape(ResourceLocation.parse(message.shapeId));
             if (shape == null) return;
             stack.set(Registration.SELECTED_SHAPE_COMPONENT.get(), new SelectedShapeData(shape));
         });